refactor: move populateMessagesDB to its own file
